pyproject.toml
[project]
name = "eps-spine-shared"
version = "0.1.0"
description = "A library to hold code shared by EPS and Spine"
authors = [
    {name = "Jolo Das", email = "jolo.das@nhs.net"},
    {name = "Anthony Brown", email = "anthony.brown8@nhs.net"},
    {name = "Matthew Popat", email = "matthew.popat1@nhs.net"},
    {name = "Phil Gee", email = "phil.gee@nhs.net"},
    {name = "Jack Spagnoli", email = "jack.spagnoli1@nhs.net"},
]
readme = "README.md"
requires-python = "==3.12.12"
dependencies = ["boto3 (>=1.42.0,<2.0.0)", "botocore (>=1.41.6,<2.0.0)", "simplejson (>=3.20.2,<4.0.0)", "python-dateutil (>=2.9.0.post0,<3.0.0)", "six (>=1.17.0,<2.0.0)"]

[tool.poetry]
packages = [{include = "eps_spine_shared", from = "src"}]

[tool.poetry.group.dev.dependencies]
black = "^25.11"
coverage = "^7.12.0"
flake8 = "^7.0"
flake8-bugbear = "^25.11"
isort = "^7.0"
pip-licenses = "^5.0"
pre-commit = "^4.0"
pytest = "^9.0"
parameterized = "^0.9.0"
freezegun = "^1.5.5"
moto = "^5.1.18"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["eps_spine_shared"]
